requirements.txt
ipaddress>=1.0.23
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
tkinter 